pyproject.toml
[project]
name = "gexa"
version = "0.1.0"
description = "GEXA - Web Search API for AI Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.3.6",
    "asyncpg>=0.30.0",
    
    # Web Crawling
    "playwright>=1.49.0",
    "beautifulsoup4>=4.12.3",
    "trafilatura>=2.0.0",
    "lxml>=5.3.0",
    
    # LLM & Embeddings
    "google-generativeai>=0.8.3",
    
    # Utilities
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    
    # Caching & Queue
    "redis>=5.2.0",
    "celery>=5.4.0",
    
    # Security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gexa"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
